Add ProcFunction model
