+++ spec/keisan/functions/proc_function_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-RSpec.describe Keisan::Function do
+RSpec.describe Keisan::Functions::ProcFunction do
   it "can be initialized from a proc" do
     function = described_class.new("test", Proc.new {|x,y| x + 3*y})
     expect(function.name).to eq "test"
